refactor(progression): separate leaderboard rendering concerns
